# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenG2P Approval Workflow Engine",
    "description": "Generic, configurable multi-stage approval workflow engine for OpenG2P modules. Caller services post artifacts; AWE resolves stages and approvers, then notifies the caller via signed webhook callbacks when state changes.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/awe/requests": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Create an approval request for a caller-owned artifact",
        "operationId": "create_request_v1_awe_requests_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRequestOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Engine refused to start the request (e.g. malformed context).",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-007",
                      "message": "Engine could not start request"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Policy key has no version in `active` status \u2014 no flow can be started for it.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-001",
                      "message": "No active policy for 'registry.cr.v1'"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Search requests by artifact reference and/or status",
        "operationId": "search_requests_v1_awe_requests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Type"
            }
          },
          {
            "name": "artifact_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestOut"
                  },
                  "title": "Response Search Requests V1 Awe Requests Get"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/awe/requests/{request_id}": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Fetch an approval request by id",
        "operationId": "get_request_v1_awe_requests__request_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No approval request with the given id.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-003",
                      "message": "Request not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/awe/requests/{request_id}/cancel": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Cancel an in-flight approval request (admin only)",
        "operationId": "cancel_request_v1_awe_requests__request_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks the `AWE_ADMIN` role required for this mutation.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_ADMIN role"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No approval request with the given id.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-003",
                      "message": "Request not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource is in a state that disallows this transition. Examples: request is already in a terminal state (`approved`/`rejected`/`cancelled`); task is not `open`/`claimed`; policy version is not `draft` for editing; delivery has already succeeded.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-007",
                      "message": "Request is in terminal state 'approved' \u2014 cannot cancel"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/awe/requests/{request_id}/events": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Timeline of every event for a request (audit log)",
        "operationId": "request_events_v1_awe_requests__request_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response Request Events V1 Awe Requests  Request Id  Events Get"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "CancelRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          }
        },
        "type": "object",
        "title": "CancelRequest"
      },
      "CreateRequestIn": {
        "properties": {
          "policy_key": {
            "type": "string",
            "title": "Policy Key",
            "examples": [
              "registry.change_request.v1"
            ]
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type",
            "examples": [
              "registry.change_request"
            ]
          },
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id",
            "examples": [
              "cr-42"
            ]
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context",
            "examples": [
              {
                "amount": 15000,
                "district": "D1"
              }
            ]
          },
          "callback_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callback Url",
            "examples": [
              "https://registry/internal/approval-callbacks"
            ]
          },
          "callback_secret_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callback Secret Id"
          },
          "requester": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester",
            "examples": [
              "u-alice"
            ]
          }
        },
        "type": "object",
        "required": [
          "policy_key",
          "artifact_type",
          "artifact_id"
        ],
        "title": "CreateRequestIn"
      },
      "CreateRequestOut": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_stage_order": {
            "type": "integer",
            "title": "Current Stage Order"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskOut"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "status",
          "current_stage_order"
        ],
        "title": "CreateRequestOut"
      },
...
      "ErrorResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": "openg2p.awe",
            "examples": [
              "openg2p.awe"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0",
            "examples": [
              "1.0"
            ]
          },
          "responsetime": {
            "type": "string",
            "title": "Responsetime",
            "examples": [
              "2026-04-23T10:00:00.000Z"
            ]
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "responsetime",
          "errors"
        ],
        "title": "ErrorResponse"
      },
      "EventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "request_id",
          "event_type",
          "payload",
          "created_at"
        ],
        "title": "EventOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RequestOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_key": {
            "type": "string",
            "title": "Policy Key"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type"
          },
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "source_service": {
            "type": "string",
            "title": "Source Service"
          },
          "requester": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_stage_order": {
            "type": "integer",
            "title": "Current Stage Order"
          },
          "callback_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callback Url"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "policy_id",
          "policy_key",
          "policy_version",
          "artifact_type",
          "artifact_id",
          "source_service",
          "context",
          "status",
          "current_stage_order",
          "created_at",
          "updated_at"
        ],
        "title": "RequestOut"
      },
...
      "TaskOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "stage_id": {
            "type": "string",
            "title": "Stage Id"
          },
          "stage_order": {
            "type": "integer",
            "title": "Stage Order"
          },
          "assignee": {
            "type": "string",
            "title": "Assignee"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "approver"
          },
          "delegated_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delegated From"
          },
          "reassigned_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reassigned From"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "claimed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "request_id",
          "stage_id",
          "stage_order",
          "assignee",
          "status",
          "created_at"
        ],
        "title": "TaskOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}